Added a feature where when the intern hits a laptop text appears on screen. In the future that text should be randomized letters as if mashing on keyboard.
Also fixed a bug where the counter that allows object interaction didn't reset. I added a failsafe so that the counter is reset when it hits another object. As it was, the problem with if counter > 255 then counter = 0 logic is that unity skips extra steps, so it hits 251, it knows, the next number is over 255, and it doesn't reset the counter because it stopped counting already. i dunno its weird.